Making it so only superadmins can delete all report

+++ application/helpers/admin.php
@@ -122,7 +122,11 @@ class admin_Core {
 
 		$menu .= ($this_sub_page == "upload") ? Kohana::lang('ui_main.upload_reports') : "<a href=\"".url::base()."admin/reports/upload\">".Kohana::lang('ui_main.upload_reports')."</a>";
 
-		$menu .= ($this_sub_page == "deleteall") ? Kohana::lang('ui_admin.delete_all') : "<a href=\"".url::base()."admin/reports/deleteall\">".Kohana::lang('ui_admin.delete_all')."</a>";
+		//only super admins have access to this
+		if (Auth::instance()->get_user()->has( ORM::factory('role','superadmin') ))
+		{
+			$menu .= ($this_sub_page == "deleteall") ? Kohana::lang('ui_admin.delete_all') : "<a href=\"".url::base()."admin/reports/deleteall\">".Kohana::lang('ui_admin.delete_all')."</a>";
+		}
 
 		echo $menu;
